+++ src/gen-bridge/GenBridge/Dhall/ExprViews.hs
@@ -1,0 +1,10 @@
+module GenBridge.Dhall.ExprViews where
+
+import Dhall.Core
+import Dhall.Map qualified as Map
+import Utils.Prelude
+
+recordField :: Text -> Expr s a -> Maybe (Expr s a)
+recordField fieldName = \case
+  RecordLit fields -> recordFieldValue <$> Map.lookup fieldName fields
+  _ -> Nothing