+++ src/gen-bridge/GenBridge/Dhall/Orphans.hs
@@ -1,0 +1,6 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module GenBridge.Dhall.Orphans where
+
+import GenBridge.Dhall.Orphans.NonEmpty ()
+import GenBridge.Dhall.Orphans.Path ()